+++ R/casting.R
@@ -31,17 +31,19 @@ detectColsToCast <- function(table, cols) {
   vals <- intersect(names(colTypes), names(cols))
   origColType <- unlist(cols[vals])
   newColType <- unlist(colTypes[vals])
-  # will consider integer and numeric as interchangeable
-  origColType <- purrr::map_chr(origColType, ~ dplyr::case_when(
-    .x == "integer" ~ "integerish",
-    .x == "numeric" ~ "integerish",
-    TRUE ~ .x
-  ))
-  newColType <- purrr::map_chr(newColType, ~ dplyr::case_when(
-    .x == "integer" ~ "integerish",
-    .x == "numeric" ~ "integerish",
-    TRUE ~ .x
-  ))
+  # if no local cdm will consider integer and numeric interchangeable
+  if (!"local_cdm" %in% class(cdmSource(table))) {
+    origColType <- purrr::map_chr(origColType, ~ dplyr::case_when(
+      .x == "integer" ~ "integerish",
+      .x == "numeric" ~ "integerish",
+      TRUE ~ .x
+    ))
+    newColType <- purrr::map_chr(newColType, ~ dplyr::case_when(
+      .x == "integer" ~ "integerish",
+      .x == "numeric" ~ "integerish",
+      TRUE ~ .x
+    ))
+  }
   differentValues <- vals[origColType != newColType]
   colsToCast <- list(
     "new" = cols[differentValues], "old" = colTypes[differentValues]